Splitting up tests to be more maintainable
