+++ tests/src/ApiLambdaAppTests.ts
@@ -1,13 +1,9 @@
 import { Expect, AsyncTest, TestFixture, TestCase } from "alsatian"
 import { Container } from "inversify"
 
-import { ApiLambdaApp, AppConfig, ErrorInterceptor, RequestBuilder } from "../../index"
+import { ApiLambdaApp, AppConfig, RequestBuilder } from "../../index"
 
 import { TestBase } from "./TestBase"
-import { TestAuthFilter } from "./test-components/TestAuthFilter"
-import { TestAuthorizer } from "./test-components/TestAuthorizer"
-import { TestDecoratorErrorInterceptor } from "./test-components/TestDecoratorErrorInterceptor"
-import { TestErrorInterceptor } from "./test-components/TestErrorInterceptor"
 
 @TestFixture()
 export class ApiLambdaAppTests extends TestBase {
@@ -56,315 +52,5 @@ export class ApiLambdaAppTests extends TestBase {
         )
 
         Expect(response.statusCode).toEqual(200)
-    }
-
-    @TestCase(null)
-    @TestCase(undefined)
-    @AsyncTest()
-    public async when_invalid_api_error_interceptor_is_passed_to_app_then_throws_an_error(invalidErrorInterceptor: ErrorInterceptor) {
-        Expect(() => this.app.middlewareRegistry.addErrorInterceptor(invalidErrorInterceptor))
-            .toThrow()
-    }
-
-    @TestCase({ controller: "TestController" })
-    @TestCase({ endpoint: "TestController::raiseError" })
-    @AsyncTest()
-    public async when_api_error_interceptor_is_configured_and_it_throws_an_error_then_interceptor_is_invoked(testCase: any) {
-        let errorInterceptor = new TestErrorInterceptor(testCase.endpoint, testCase.controller)
-
-        this.app.middlewareRegistry.addErrorInterceptor(errorInterceptor)
-
-        await this.sendRequest(
-            RequestBuilder.get("/test/raise-error").build()
-        )
-
-        Expect(errorInterceptor.wasInvoked).toBeTruthy()
-    }
-
-    @AsyncTest()
-    public async when_api_controller_decorator_error_interceptor_is_present_and_it_throws_an_error_then_interceptor_is_invoked() {
-        TestDecoratorErrorInterceptor.wasInvoked = false
-
-        await this.sendRequest(
-            RequestBuilder.get("/test/controller-ei-decorator").build()
-        )
-
-        Expect(TestDecoratorErrorInterceptor.wasInvoked).toBeTruthy()
-    }
-
-    @AsyncTest()
-    public async when_api_endpoint_decorator_error_interceptor_is_present_and_it_throws_an_error_then_interceptor_is_invoked() {
-        TestDecoratorErrorInterceptor.wasInvoked = false
-
-        await this.sendRequest(
-            RequestBuilder.get("/test/ei-decorator").build()
-        )
-
-        Expect(TestDecoratorErrorInterceptor.wasInvoked).toBeTruthy()
-    }
-
-    @AsyncTest()
-    public async when_api_error_interceptor_is_configured_and_no_error_is_thrown_then_interceptor_is_not_invoked() {
-        let errorInterceptor = new TestErrorInterceptor(null, "TestController")
-
-        this.app.middlewareRegistry.addErrorInterceptor(errorInterceptor)
-
-        await this.sendRequest(
-            RequestBuilder.get("/test").build()
-        )
-
-        Expect(errorInterceptor.wasInvoked).toBe(false)
-    }
-
-    @AsyncTest()
-    public async when_api_error_interceptor_is_invoked_and_no_response_is_returned_by_interceptor_then_original_error_is_returned() {
-        this.app.middlewareRegistry.addErrorInterceptor(
-            new TestErrorInterceptor("TestController::raiseError")
-        )
-
-        let response = await this.sendRequest(
-            RequestBuilder.get("/test/raise-error").build()
-        )
-
-        Expect(response.statusCode).toEqual(500)
-        Expect(response.body).toEqual("{\"error\":\"all I do is throw an error\"}")
-    }
-
-    @AsyncTest()
-    public async when_api_error_interceptor_is_invoked_and_response_is_returned_by_interceptor_then_interceptor_response_is_returned() {
-        this.app.middlewareRegistry.addErrorInterceptor(
-            new TestErrorInterceptor("TestController::raiseError", null, true)
-        )
-
-        let response = await this.sendRequest(
-            RequestBuilder.get("/test/raise-error").build()
-        )
-
-        Expect(response.statusCode).toEqual(200)
-        Expect(response.body).toEqual("interceptor return value")
-    }
-
-    @TestCase("/test/raise-error")
-    @TestCase("/test/methods/raise-error")
-    @AsyncTest()
-    public async when_global_api_error_interceptor_is_invoked_and_error_is_thrown_in_any_endpoint_then_interceptor_is_invoked(path: string) {
-        let errorInterceptor = new TestErrorInterceptor("*", null)
-
-        this.app.middlewareRegistry.addErrorInterceptor(errorInterceptor)
-
-        await this.sendRequest(
-            RequestBuilder.get(path).build()
-        )
-
-        Expect(errorInterceptor.wasInvoked).toBeTruthy()
-    }
-
-    @TestCase(null)
-    @TestCase(undefined)
-    @AsyncTest()
-    public async when_invalid_api_auth_filter_is_passed_to_app_then_throws_an_error(invalidAuthFilter: TestAuthFilter) {
-        Expect(() => this.app.middlewareRegistry.addAuthFilter(invalidAuthFilter))
-            .toThrow()
-    }
-
-    @AsyncTest()
-    public async when_api_auth_filter_configured_and_request_is_made_then_filter_is_invoked() {
-        let authFilter = new TestAuthFilter("luke", "vaderismydad")
-
-        this.app.middlewareRegistry.addAuthFilter(authFilter)
-
-        await this.sendRequest(
-            RequestBuilder.get("/test")
-                .basicAuth("luke", "vaderismydad")
-                .build()
-        )
-
-        Expect(authFilter.wasInvoked).toBeTruthy()
-    }
-
-    @AsyncTest()
-    public async when_api_auth_filter_configured_and_request_is_made_then_credentials_are_passed_to_filter() {
-        let authFilter = new TestAuthFilter("luke", "vaderismydad")
-
-        this.app.middlewareRegistry.addAuthFilter(authFilter)
-
-        await this.sendRequest(
-            RequestBuilder.get("/test")
-                .basicAuth("luke", "vaderismydad")
-                .build()
-        )
-
-        Expect(authFilter.passedCredentials.username).toEqual("luke")
-        Expect(authFilter.passedCredentials.password).toEqual("vaderismydad")
-    }
-
-    @TestCase("vaderismydad", 200)
-    @TestCase("whoismydad?", 401)
-    @AsyncTest()
-    public async when_api_auth_filter_configured_and_credentials_passed_then_valid_endpoint_request_returns_correct_status(password: string, expectedStatus: number) {
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("luke", "vaderismydad")
-        )
-
-        let response = await this.sendRequest(
-            RequestBuilder.get("/test")
-                .basicAuth("luke", password)
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(expectedStatus)
-    }
-
-    @AsyncTest()
-    public async when_api_auth_filter_configured_and_no_credentials_passed_then_valid_endpoint_request_returns_401_unauthorized() {
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("luke", "vaderismydad")
-        )
-
-        let response = await this.sendRequest(
-            RequestBuilder.get("/test")
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(401)
-    }
-
-    @AsyncTest()
-    public async when_api_auth_filter_configured_and_filter_throws_error_then_request_returns_500_server_error() {
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("luke", "vaderismydad", true)
-        )
-
-        let response = await this.sendRequest(
-            RequestBuilder.get("/test")
-                .basicAuth("luke", "vaderismydad")
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(500)
-    }
-
-    @AsyncTest()
-    public async when_api_auth_filter_configured_and_credentials_passed_then_endpoint_is_passed_current_principal() {
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("stoat", "ihavenoideawhatiam")
-        )
-
-        let response = await this.sendRequest(
-            RequestBuilder.get("/test/user-test")
-                .basicAuth("stoat", "ihavenoideawhatiam")
-                .build()
-        )
-
-        Expect(response.body).toEqual("stoat")
-    }
-
-    @TestCase(null)
-    @TestCase(undefined)
-    @AsyncTest()
-    public async when_invalid_api_authorizer_is_passed_to_app_then_throws_an_error(invalidAuthorizer: TestAuthorizer) {
-        Expect(() => this.app.middlewareRegistry.addAuthorizer(invalidAuthorizer))
-            .toThrow()
-    }
-
-    @AsyncTest()
-    public async when_api_authorizer_configured_and_no_roles_declared_then_unprotected_resource_returns_200_ok() {
-        this.app.middlewareRegistry.addAuthorizer(new TestAuthorizer())
-
-        let response = await this.sendRequest(
-            RequestBuilder.get("/test")
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(200)
-    }
-
-    @TestCase("/test/restricted")
-    @TestCase("/test-restricted")
-    @AsyncTest()
-    public async when_api_authorizer_not_configured_and_roles_declared_then_protected_resource_returns_500_server_error(path: string) {
-        let response = await this.sendRequest(
-            RequestBuilder.get(path)
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(500)
-    }
-
-    @TestCase({ path: "/test/restricted", userRoles: ["SPECIAL_USER"] })
-    @TestCase({ path: "/test-restricted", userRoles: ["SPECIAL_USER"] })
-    @TestCase({ path: "/test-restricted", userRoles: ["SUPER_SPECIAL_USER"] })
-    @AsyncTest()
-    public async when_api_authorizer_configured_with_roles_declared_and_user_is_authorized_then_valid_request_returns_200_ok(testCase: any) {
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("stoat", "ihavenoideawhatiam", false, testCase.userRoles)
-        )
-        this.app.middlewareRegistry.addAuthorizer(new TestAuthorizer())
-
-        let response = await this.sendRequest(
-            RequestBuilder.get(testCase.path)
-                .basicAuth("stoat", "ihavenoideawhatiam")
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(200)
-    }
-
-    @TestCase({ path: "/test/restricted", userRoles: ["SPECIAL_USER"] })
-    @TestCase({ path: "/test-restricted", userRoles: ["SPECIAL_USER"] })
-    @TestCase({ path: "/test-restricted", userRoles: ["SUPER_SPECIAL_USER"] })
-    @AsyncTest()
-    public async when_api_authorizer_configured_with_roles_declared_and_user_is_authorized_and_authorizer_throws_error_then_valid_request_returns_500_server_error(testCase: any) {
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("stoat", "ihavenoideawhatiam", false, testCase.userRoles)
-        )
-        this.app.middlewareRegistry.addAuthorizer(new TestAuthorizer(true))
-
-        let response = await this.sendRequest(
-            RequestBuilder.get(testCase.path)
-                .basicAuth("stoat", "ihavenoideawhatiam")
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(500)
-    }
-
-    @TestCase("/test/restricted")
-    @TestCase("/test-restricted")
-    @AsyncTest()
-    public async when_api_authorizer_configured_with_roles_declared_and_user_is_not_authorized_then_valid_request_returns_403_forbidden(path: string) {
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("stoat", "ihavenoideawhatiam", false, ["ANOTHER_THING"])
-        )
-        this.app.middlewareRegistry.addAuthorizer(new TestAuthorizer())
-
-        let response = await this.sendRequest(
-            RequestBuilder.get(path)
-                .basicAuth("stoat", "ihavenoideawhatiam")
-                .build()
-        )
-
-        Expect(response.statusCode).toEqual(403)
-    }
-
-    @TestCase("/test/restricted")
-    @TestCase("/test-restricted")
-    @AsyncTest()
-    public async when_api_authorizer_configured_and_role_declared_then_principle_and_role_are_passed_to_authorizer(path: string) {
-        let authorizer = new TestAuthorizer()
-
-        this.app.middlewareRegistry.addAuthFilter(
-            new TestAuthFilter("stoat", "ihavenoideawhatiam")
-        )
-        this.app.middlewareRegistry.addAuthorizer(authorizer)
-
-        await this.sendRequest(
-            RequestBuilder.get(path)
-                .basicAuth("stoat", "ihavenoideawhatiam")
-                .build()
-        )
-
-        Expect(authorizer.principalPassed.name).toEqual("stoat")
-        Expect(authorizer.rolePassed).toEqual("SPECIAL_USER")
     }
 }
